+++ code/qcommon/con_log.c
@@ -48,7 +48,7 @@ static unsigned int CON_LogFree( void )
 
 unsigned int CON_LogWrite( const char *in )
 {
-	unsigned int length = strlen( in );
+	unsigned int length = (unsigned int)strlen( in );
 	unsigned int firstChunk;
 	unsigned int secondChunk;
